Consolidate focus behavior into concern
