+++ .env.test
@@ -6,5 +6,3 @@ COVERAGE=false
 
 # Set to true to run test suite serially
 DISABLE_PARALLELIZATION=false
-
-E2E_BROWSER=headless_chrome
