+++ app/controllers/concerns/accountable_resource.rb
@@ -2,6 +2,8 @@ module AccountableResource
   extend ActiveSupport::Concern
 
   included do
+    include ScrollFocusable
+
     layout :with_sidebar
     before_action :set_account, only: [ :show, :edit, :update, :destroy ]
     before_action :set_link_token, only: :new
@@ -25,19 +27,9 @@ module AccountableResource
     @q = params.fetch(:q, {}).permit(:search)
     entries = @account.entries.search(@q).reverse_chronological
 
-    if params[:focused_entry_id].present?
-      @focused_entry = entries.find_by(id: params[:focused_entry_id])
-      position = entries.pluck(:id).index(params[:focused_entry_id])
+    set_focused_record(entries, params[:focused_record_id])
 
-      if position.present?
-        focused_page = (position / (params[:per_page] || 10)) + 1
-        if params[:page]&.to_i != focused_page
-          return redirect_to account_path(@account, page: focused_page, focused_entry_id: params[:focused_entry_id])
-        end
-      end
-    end
-
-    @pagy, @entries = pagy(entries, limit: params[:per_page] || "10", params: ->(params) { params.except(:focused_entry_id) })
+    @pagy, @entries = pagy(entries, limit: params[:per_page] || "10", params: ->(params) { params.except(:focused_record_id) })
   end
 
   def edit
