+++ app/controllers/transactions_controller.rb
@@ -1,4 +1,6 @@
 class TransactionsController < ApplicationController
+  include ScrollFocusable
+
   layout :with_sidebar
 
   before_action :store_params!, only: :index
@@ -7,21 +9,13 @@ class TransactionsController < ApplicationController
     @q = search_params
     search_query = Current.family.transactions.search(@q).reverse_chronological
 
-    if params[:focused_entry_id].present?
-      @focused_entry = search_query.find_by(id: params[:focused_entry_id])
-      position = search_query.pluck(:id).index(params[:focused_entry_id])
+    set_focused_record(search_query, params[:focused_record_id], default_per_page: 50)
 
-      if position.present?
-        focused_page = (position / (params[:per_page] || 10).to_i) + 1
-        if params[:page]&.to_i != focused_page
-          return redirect_to transactions_path(page: focused_page, focused_entry_id: params[:focused_entry_id])
-        else
-          params.delete(:focused_entry_id)
-        end
-      end
-    end
-
-    @pagy, @transaction_entries = pagy(search_query, limit: params[:per_page].presence || default_params[:per_page], params: ->(params) { params.except(:focused_entry_id) })
+    @pagy, @transaction_entries = pagy(
+      search_query, 
+      limit: params[:per_page].presence || default_params[:per_page], 
+      params: ->(params) { params.except(:focused_record_id) }
+    )
 
     totals_query = search_query.incomes_and_expenses
     family_currency = Current.family.currency
