+++ app/views/accounts/show/_activity.html.erb
@@ -1,7 +1,7 @@
 <%# locals: (account:) %>
 
 <%= turbo_frame_tag dom_id(account, "entries") do %>
-  <div class="bg-white p-5 border border-alpha-black-25 rounded-xl shadow-xs" data-controller="focus-entry" data-focus-entry-id-value="<%= @focused_entry ? dom_id(@focused_entry) : nil %>">
+  <div class="bg-white p-5 border border-alpha-black-25 rounded-xl shadow-xs" data-controller="focus-record" data-focus-record-id-value="<%= @focused_record ? dom_id(@focused_record) : nil %>">
     <div class="flex items-center justify-between mb-4">
       <%= tag.h2 t(".title"), class: "font-medium text-lg" %>
       <% unless @account.plaid_account_id.present? %>
